Fix changelog display and linking issues
+ The CHANGELOG used a `:heart:` instead of a unicode character; this only
works in GFM, and so did not translate to documentation. Updated the header
to use the unicode heart instead.
+ The CHANGELOG has numerous referenced issues but github does not link to
the filed issues natively. Additionally, the CHANGELOG display was broken
in GitBook due to a mis-parsing of the issues as headers. Here we turned
each issue into an explicit link to the upstream issue tracker and fixed
both problems at once.